integrate command bus and event bus extension
